add default arguments to types; remove inconsistent duck typing
"crs" and "id" arguments are now optional. "crs" has been removed from
the GeoJSON spec, while "id" was always non-compulsory.

duck typing for Feature, FeatureCollection, and GeometryCollection were
workarounds for the orginal karta code, and look messy here. Serializer
requires one of the provided named tuples as input, and it is the
responsibility of the caller to convert other types.