implements antimeridian cutting for polygons
Handles polygon holes by testing all external/internal ring combinations
(inefficient, but a more advanced approach feels out of scope for now)
to determine which (possibly split) hole fits in which polygon.

To do so, implements a basic winding number scheme[0], and checks for
any internal ring vertex contained within each external ring. In a
cartesian setting, this is good enough, although with true geographical
coordinates this is not strictly correct. I justify this based on
RFC7946, which states:

    A line between two positions is a straight Cartesian line, the
    shortest line between those two points in the coordinate reference
    system (see Section 4).

[0] Algorithm copied from http://geomalgorithms.com/a03-_inclusion.html