give Deserializer a cleaner public API
Now, similar to Serializer, the Deserializer is created and fed input
via load_json (for raw JSON) or read_json (for file objects or
filenames), yielding parsed ("deserialized") GeoJSON tuples.

As a shortcut, the Deserializer instance can be called with the input
(filename, file object, or raw JSON), and it will determine the proper
approach and return a GeoJSON tuple.